chore: Move files related to mining state to `state/mining/`
Contributes to balancing `state/`.
