+++ neptune-core/src/application/control/main_loop.rs
@@ -56,10 +56,10 @@ use crate::models::peer::peer_info::PeerInfo;
 use crate::models::peer::transaction_notification::TransactionNotification;
 use crate::models::peer::PeerSynchronizationState;
 use crate::models::proof_abstractions::tasm::program::TritonVmProofJobOptions;
-use crate::state::block_proposal::BlockProposal;
 use crate::state::mempool::mempool_update_job::MempoolUpdateJob;
 use crate::state::mempool::mempool_update_job_result::MempoolUpdateJobResult;
 use crate::state::mempool::upgrade_priority::UpgradePriority;
+use crate::state::mining::block_proposal::BlockProposal;
 use crate::state::networking_state::SyncAnchor;
 use crate::state::tx_proving_capability::TxProvingCapability;
 use crate::state::GlobalState;
