+++ neptune-core/src/application/control/peer_loop.rs
@@ -56,8 +56,8 @@ use crate::models::peer::PositivePeerSanction;
 use crate::models::peer::SyncChallenge;
 use crate::models::proof_abstractions::mast_hash::MastHash;
 use crate::models::proof_abstractions::timestamp::Timestamp;
-use crate::state::block_proposal::BlockProposalRejectError;
 use crate::state::mempool::MEMPOOL_TX_THRESHOLD_AGE_IN_SECS;
+use crate::state::mining::block_proposal::BlockProposalRejectError;
 use crate::state::GlobalState;
 use crate::state::GlobalStateLock;
 use crate::util_types::mutator_set::removal_record::RemovalRecordValidityError;
