+++ neptune-core/src/application/rpc/server.rs
@@ -97,8 +97,8 @@ use crate::models::peer::peer_info::PeerInfo;
 use crate::models::peer::InstanceId;
 use crate::models::peer::PeerStanding;
 use crate::models::proof_abstractions::timestamp::Timestamp;
-use crate::state::mining_state::MAX_NUM_EXPORTED_BLOCK_PROPOSAL_STORED;
-use crate::state::mining_status::MiningStatus;
+use crate::state::mining::mining_state::MAX_NUM_EXPORTED_BLOCK_PROPOSAL_STORED;
+use crate::state::mining::mining_status::MiningStatus;
 use crate::state::transaction_details::TransactionDetails;
 use crate::state::transaction_kernel_id::TransactionKernelId;
 use crate::state::tx_creation_artifacts::TxCreationArtifacts;
@@ -4994,7 +4994,7 @@ mod tests {
         use crate::models::blockchain::block::pow::Pow;
         use crate::models::blockchain::block::BlockProof;
         use crate::models::blockchain::transaction::validity::neptune_proof::NeptuneProof;
-        use crate::state::block_proposal::BlockProposal;
+        use crate::state::mining::block_proposal::BlockProposal;
         use crate::state::wallet::address::generation_address::GenerationReceivingAddress;
         use crate::state::wallet::address::KeyType;
         use crate::tests::shared::blocks::fake_deterministic_successor;
