+++ neptune-core/src/state/mining/mining_status.rs
