+++ neptune-core/src/state/mining/mod.rs
@@ -1,0 +1,3 @@
+pub mod block_proposal;
+pub mod mining_state;
+pub mod mining_status;
