+++ neptune-core/src/state/mod.rs
@@ -1,10 +1,8 @@
 pub mod archival_state;
-pub mod block_proposal;
 pub mod blockchain_state;
 pub mod light_state;
 pub mod mempool;
-pub mod mining_state;
-pub mod mining_status;
+pub mod mining;
 pub mod networking_state;
 pub mod shared;
 pub(crate) mod transaction_details;
@@ -30,16 +28,16 @@ use std::time::SystemTime;
 use anyhow::bail;
 use anyhow::ensure;
 use anyhow::Result;
-use block_proposal::BlockProposal;
 use blockchain_state::BlockchainArchivalState;
 use blockchain_state::BlockchainState;
 use itertools::Itertools;
 use light_state::LightState;
 use mempool::Mempool;
-use mining_state::MiningState;
-use mining_status::ComposingWorkInfo;
-use mining_status::GuessingWorkInfo;
-use mining_status::MiningStatus;
+use mining::block_proposal::BlockProposal;
+use mining::mining_state::MiningState;
+use mining::mining_status::ComposingWorkInfo;
+use mining::mining_status::GuessingWorkInfo;
+use mining::mining_status::MiningStatus;
 use networking_state::NetworkingState;
 use num_traits::CheckedSub;
 use num_traits::Zero;
@@ -93,9 +91,9 @@ use crate::models::peer::SyncChallenge;
 use crate::models::peer::SyncChallengeResponse;
 use crate::models::peer::SYNC_CHALLENGE_POW_WITNESS_LENGTH;
 use crate::models::proof_abstractions::timestamp::Timestamp;
-use crate::state::block_proposal::BlockProposalRejectError;
 use crate::state::mempool::mempool_update_job::MempoolUpdateJob;
 use crate::state::mempool::upgrade_priority::UpgradePriority;
+use crate::state::mining::block_proposal::BlockProposalRejectError;
 use crate::state::wallet::expected_utxo::ExpectedUtxo;
 use crate::state::wallet::expected_utxo::UtxoNotifier;
 use crate::state::wallet::monitored_utxo::MonitoredUtxo;
