+++ neptune-dashboard/src/overview_screen.rs
@@ -13,7 +13,7 @@ use neptune_cash::application::rpc::server::RPCClient;
 use neptune_cash::models::blockchain::block::block_header::BlockHeader;
 use neptune_cash::models::blockchain::block::block_height::BlockHeight;
 use neptune_cash::models::blockchain::type_scripts::native_currency_amount::NativeCurrencyAmount;
-use neptune_cash::state::mining_status::MiningStatus;
+use neptune_cash::state::mining::mining_status::MiningStatus;
 use neptune_cash::state::tx_proving_capability::TxProvingCapability;
 use ratatui::layout::Margin;
 use ratatui::layout::Rect;
